refactor(autofix): split scope routing and test writer modules
